+++ src/database/sql/create_session.sql
@@ -1,5 +1,8 @@
+-- DROP TABLE IF EXISTS session
 CREATE TABLE IF NOT EXISTS session (
-    pid INTEGER NOT NULL,
+    pid BIGINT NOT NULL,
     cmd TEXT    NOT NULL
 );
+
+-- DROP INDEX IF EXISTS session_pid_index
 CREATE INDEX IF NOT EXISTS session_pid_index ON session(pid);
